+++ src/ResourcePathMap.java
@@ -1,25 +1,25 @@
 import java.util.*;
 
 public class ResourcePathMap {
-    private PrefixTree<ResourcePath, Integer> resourceTree;
+    private PrefixTree<ResourcePath, Long> resourceTree;
 
     public ResourcePathMap() {
-        this.resourceTree = new PrefixTree<ResourcePath, Integer>();
+        this.resourceTree = new PrefixTree<ResourcePath, Long>();
     }
 
     public boolean contains(ResourcePath path) {
         return this.resourceTree.contains(path);
     }
 
-    public void put(ResourcePath path, Integer value) {
+    public void put(ResourcePath path, Long value) {
         this.resourceTree.put(path, value);
     }
 
-    public Integer get(ResourcePath path) {
+    public Long get(ResourcePath path) {
         return this.resourceTree.get(path);
     }
 
-    public PrefixTree<ResourcePath, Integer> getResourceTree() {
+    public PrefixTree<ResourcePath, Long> getResourceTree() {
         return this.resourceTree;
     }
 }
